+++ R/mirtCAT-package.R
@@ -27,5 +27,6 @@
 #' @import mirt shiny methods lattice Rcpp
 #' @importFrom stats approx na.omit pnorm qnorm reshape runif spline
 #' @importFrom lpSolve lp
+#' @importFrom pbapply pblapply
 #' @keywords package
 NULL